requirements.txt
requests>=2.28.0
pymongo>=4.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
prometheus-fastapi-instrumentator>=7.0.0
httpx>=0.24.0
langfuse
PyJWT
argon2-cffi
email-validator
python-multipart>=2.0.0
